% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_01_airData.R
\docType{class}
\name{airData,-class}
\alias{airData,-class}
\alias{airData}
\alias{airData-class}
\title{S4 Class airData}
\description{
Used to provide a single object class for various sources of sensor data.
For instance, CSV data obtained from madavi.de or luftdaten.info has a
different format even if the same device was used.
}
\section{Slots}{

\describe{
\item{\code{data}}{List of data frames containing all raw data.}

\item{\code{layout}}{Data frame with two columns, \code{obj}, \code{data},
      and \code{time},
and rows named after the type of data measured (e.g., \code{PM10}, \code{PM2_5},
\code{temperature}, \code{humidity}). For each type, it defines in which
data frame from the \code{data} slot the actual data can be found, and how the
column with the time data is called.}
}}

\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be
the case), the contructor function \code{airData(...)} can be used instead of
\code{new("airData", ...)}.
}

\keyword{classes}
