% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_get_set_airData.R
\docType{methods}
\name{air_data}
\alias{air_data}
\alias{air_data,airData-method}
\alias{air_data,-methods}
\alias{air_data<-}
\alias{air_data<-,airData-method}
\alias{air_data<-,-methods}
\alias{air_layout}
\alias{air_layout,airData-method}
\alias{air_layout,-methods}
\alias{air_layout<-}
\alias{air_layout<-,airData-method}
\alias{air_layout<-,-methods}
\alias{layout_df}
\alias{layout_df,airData-method}
\alias{layout_df,-methods}
\alias{[,airData-method}
\alias{[,-methods}
\alias{[,airData,ANY,ANY-method}
\alias{[<-,airData-method}
\alias{[<-,-methods}
\alias{[<-,airData,ANY,ANY,ANY-method}
\alias{[[,airData-method}
\alias{[[,-methods}
\alias{[[,airData,ANY-method}
\alias{[[<-,airData-method}
\alias{[[<-,-methods}
\alias{[[<-,airData,ANY,ANY-method}
\title{Getter and setter methods for airData objects}
\usage{
air_data(x, ...)

\S4method{air_data}{airData}(x)

air_data(x) <- value

\S4method{air_data}{airData}(x) <- value

air_layout(x, value)

\S4method{air_layout}{airData}(x, value)

air_layout(x) <- value

\S4method{air_layout}{airData}(x) <- value

layout_df(x, time = c(), ...)

\S4method{layout_df}{airData}(x, time = c())

\S4method{[}{airData}(x, i, j, time = c())

\S4method{[}{airData}(x, i, j, time = c()) <- value

\S4method{[[}{airData}(x, i, time = c())

\S4method{[[}{airData}(x, i, time = c()) <- value
}
\arguments{
\item{x}{An object of class \code{\link[paRticulates:airData-class]{airData}}.}

\item{...}{Additional arguments for the generics.}

\item{value}{The new value to replace the current with.}

\item{time}{An optional vector with two elements, start and end time, either in character
format supported by \code{\link[as.POSIXct]{as.POSIXct}}, or in POSIXct format.}

\item{i}{Defines the row selector (\code{[}) or the name to match (\code{[[}).}

\item{j}{Defines the column selector.}
}
\description{
\itemize{
  \item{\code{air_data()}/\code{air_data()<-} }{Get/set the \code{data} slot.}
  \item{\code{air_layout()}/\code{air_layout()<-} }{Get/set the \code{layout} slot.}
  \item{\code{layout_df()} }{Returns a single data frame as described by the \code{layout} slot.}
  \item{\code{[}/\code{[[} }{Can be used as a shortcut to directly access the data as described by the \code{layout} slot.}
}
}
