% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_aq_hourly.R
\docType{methods}
\name{aq_hourly}
\alias{aq_hourly}
\alias{aq_hourly,airData-method}
\alias{aq_hourly,-methods}
\title{Get hourly means and air quality indices}
\usage{
aq_hourly(x, time = c(), ...)

\S4method{aq_hourly}{airData}(x, time = c(), calc = c("CAQI", "EAQI"))
}
\arguments{
\item{x}{An object of class \code{\link[paRticulates:airData-class]{airData}}.}

\item{time}{An optional vector with two elements, start and end time, either in character
format supported by \code{\link[as.POSIXct]{as.POSIXct}}, or in POSIXct format.}

\item{calc}{Character vector of air quality indices to calculate.
Currently, "CAQI" and "EAQI" are supported.}
}
\value{
A data frame as returned by \code{\link[paRticulates:layout_df]{layout_df}},
      but reduced to hourly means.
}
\description{
The hour value is to be interpreted as the starting point of measurement until the next full hour.
this means, if you are looking for the past hour, you need to add 1 to the hour values.
}
