% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI_US_PM.R
\name{AQI_US_PM}
\alias{AQI_US_PM}
\title{Calcualte particulate matter sub-index of the US Air Quality Index (AQI)}
\usage{
AQI_US_PM(x, norm = "PM10", raw = FALSE, latex_cellcolors = TRUE)
}
\arguments{
\item{x}{A numeric vector of particulate matter measurements (µg/m³).}

\item{norm}{One value of either "PM10" or "PM2.5",
      setting the PM size of the given values of \code{x}.}

\item{raw}{Logical, if true returns only the AQI_US raw values.}

\item{latex_cellcolors}{Logical,
      if true the "AQI_US_color_*" columns will show LaTeX \code{\\cellcolor} code instead
of HTML colors. See \code{Colors} section.}
}
\value{
If \code{raw=TRUE},
      a numeric vector with the raw AQI_US values. Otherwise a data frame with
   four columns: raw input data, AQI_US value, AQI_US level (factor),
      recommended color code (factor).
   The column names start with "raw_", "AQI_US_", "AQI_US_level_", and "AQI_US_color_",
      and end with either "PM10" or
   "PM2_5", depending on the value of \code{norm}.
}
\description{
The function returns the 24-hour AQI grading for given PM raw values,
      as defined by the United States Environmental Protection Agency (EPA)[1].
}
\details{
The index has six levels, named "good" (best air quality), "moderate",
      "unhealthy for sensitive groups", "unhealthy", "very unhealthy", and "hazardous" (worst air quality).
Thresholds are different between PM10 and PM2.5 particles.
}
\note{
If there is no full 24-hours of data available,
      the AQI standard calculates a \code{\link[NowCast_PM]{NowCast}} value from
the last 12 hours. This is currently not implemented here,
      but available as a \code{\link[NowCast_PM]{separate function}}.
}
\section{Colors}{
 The data frame contains rows to support coloring tables or plots,
      e.g. if you use the data in
   an RMarkdown document. If \code{latex_cellcolors=FALSE},
      they provide a \code{<span>} with hexadecimal HTML color codes
   for each index value. Otherwise,
      the factor labels are in LaTeX format using the \code{\\cellcolor} command.
   This is useful, e.g.,
      if combined with the \code{kableExtra::kable} function. To use the code out of the box,
   define the following colors in your preamble (example shows usage in a YAML header):
   
\preformatted{
header-includes:
   - \definecolor{AQIUSg}{HTML}{00e400}
   - \definecolor{AQIUSm}{HTML}{ffff00}
   - \definecolor{AQIUSus}{HTML}{ff7e00}
   - \definecolor{AQIUSu}{HTML}{ff0000}
   - \definecolor{AQIUSvu}{HTML}{99004c}
   - \definecolor{AQIUSh}{HTML}{7e0023}
}
}

\examples{
AQI_US_PM(55)
}
\references{
[1] \url{https://airnow.gov/index.cfm?action=aqibasics.aqi}
}
\seealso{
\code{\link[NowCast_PM]{NowCast_PM}}, \code{\link[EAQI_PM]{EAQI_PM}},
      \code{\link[CAQI_PM]{CAQI_PM}}
}
