% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAQI_sub.R
\name{CAQI_sub}
\alias{CAQI_sub}
\title{Calcualte particulate matter sub-index of th EU Common Air Quality Index (CAQI)}
\usage{
CAQI_sub(x, norm = "PM10")
}
\arguments{
\item{x}{A numeric vector of particular matter measurements (µg/m³).}

\item{norm}{One value of either "PM10" or "PM2.5",
      setting the PM size of the given values of \code{x}.}
}
\value{
A named numeric vector.
}
\description{
The function returns the respective hourly CAQI grading.
}
\details{
The index has five levels, named "very low" (best air quality), "low", "medium", "high",
      and "very high" (worst air quality).
Thresholds are different between PM10 and PM2.5 particles.
}
\examples{
CAQI_sub(55)
}
