% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAQI_sub.R
\name{EAQI_sub}
\alias{EAQI_sub}
\title{Calcualte particulate matter sub-index of the European Air Quality Index (EAQI)}
\usage{
EAQI_sub(x, norm = "PM10")
}
\arguments{
\item{x}{A numeric vector of particulate matter measurements (µg/m³).}

\item{norm}{One value of either "PM10" or "PM2.5",
      setting the PM size of the given values of \code{x}.}
}
\value{
If \code{raw=TRUE}, a numeric vector with the raw EAQI values. Otherwise a data frame with
   four columns: raw input data, EAQI level (factor), recommended color code (factor).
   The column names start with "raw_", "EAQI_level_", and "EAQI_color_",
      and end with either "PM10" or
   "PM2_5", depending on the value of \code{norm}.
}
\description{
The function returns the respective hourly EAQI grading.
}
\details{
The index has six levels, named "good" (best air quality), "fair", "moderate", "poor",
      "very poor",
and "extremely poor" (worst air quality).
Thresholds are different between PM10 and PM2.5 particles.
}
\examples{
EAQI_sub(55)
}
