% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NowCast_PM.R
\name{NowCast_PM}
\alias{NowCast_PM}
\title{Calcualte particulate matter NowCast}
\usage{
NowCast_PM(x, norm = "PM10", raw = FALSE, latex_cellcolors = TRUE)
}
\arguments{
\item{x}{A numeric vector of length 12, being the PM raw values of the past 12 hours,
      with the last being the most recent one.}

\item{norm}{One value of either "PM10" or "PM2.5",
      setting the PM size of the given values of \code{x}.}

\item{raw}{Logical, if true returns only the NowCast raw value.}

\item{latex_cellcolors}{Logical,
      if true the "NowCast_color_*" columns will show LaTeX \code{\\cellcolor} code instead
of HTML colors. See \code{Colors} section of \code{\link[AQI_US_PM]{AQI_US_PM}}.}
}
\value{
If \code{raw=TRUE},
      a numeric vector with the raw NowCast values. Otherwise a data frame with
   four columns: NowCast raw value (weighted mean of input data), AQI raw value,
      AQI level (factor), recommended color code (factor).
   The column names start with "NowCast_raw_", "NowCast_AQI_", "NowCast_AQI_level_",
      and "NowCast_AQI_color_", and end with either "PM10" or
   "PM2_5", depending on the value of \code{norm}.
}
\description{
The function returns the weighted mean of exctly 12 given PM raw values,
      as defined by the United States Environmental Protection Agency (EPA)[1].
}
\examples{
NowCast_PM(c(50, 80, 75, 90, 82, 53, 64, 74, 21, 10, 16, 13))
}
\references{
[1] \url{https://en.wikipedia.org/wiki/NowCast_(air_quality_index)}
}
