% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paRticulates-package.R
\docType{package}
\name{paRticulates-package}
\alias{paRticulates}
\alias{paRticulates-package}
\title{\packageTitle{paRticulates}}
\description{
\packageDescription{paRticulates}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab paRticulates\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2020-01-15\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=paRticulates\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=paRticulates}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/paRticulates/issues}
}

}
\author{
\packageAuthor{paRticulates}

Maintainer: \packageMaintainer{paRticulates}
}
