% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_plot.airData.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,airData,missing-method}
\title{Plot air quality data from an \code{airData} object}
\usage{
plot(x, y, ...)

\S4method{plot}{airData,missing}(
  x,
  main = "Particulates",
  start = NA,
  end = NA,
  sub = paste0("Date: "),
  xlab = "Time of day",
  ylab1 = "Humidity",
  ylab2 = "Particulates (µg/m³)",
  zlab = "Temperature",
  legend = c("Humid", "Temp", "PM10", "PM2.5"),
  colors = c("blue", "red", "darkgreen", "green"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{airData}, like returned by read.madavi().}

\item{...}{Additional options applied to \code{plot} and \code{lines}.}

\item{main}{Title of the plot.}

\item{start}{Start time in POSIXct format,
      or as a character string supported by \code{\link[as.POSIXct]{as.POSIXct}}. If missing all data is plotted.}

\item{end}{End time in POSIXct format,
      or as a character string supported by \code{\link[as.POSIXct]{as.POSIXct}}. If missing all data is plotted.}

\item{sub}{Subtitle}

\item{xlab}{A title for the x axis (time).}

\item{ylab1}{A title for the upper y axis (humidity).}

\item{ylab2}{A title for the lower y axis (fine dust).}

\item{zlab}{A title for the upper z axis (temperature).}

\item{legend}{Text for the plot legend describing humidity, temperature, PM10, and PM2.5.}

\item{colors}{Four color definitions used to plot humidity, temperature, PM10, and PM2.5,
      respectively.}
}
\description{
Plots objects of class \code{airData}.
}
\examples{
\dontrun{
x <- read.madavi(file.path("~","fine_dust_data"))
plot(
  x,
  start="2019-12-10 09:00:00",
  end="2019-12-10 22:00:00"
)
}
}
\keyword{methods}
\keyword{plot}
