% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.airData.R
\name{plot.airData}
\alias{plot.airData}
\title{Plot air quality data from an airData object}
\usage{
\method{plot}{airData}(
  data,
  main = "Feinstaubkonzentration",
  start = NA,
  end = NA,
  sub = paste0("Datum: ", unique(format(data[["TimeLocal"]], "\%d.\%m.\%Y"))),
  xlab = "Tageszeit",
  ylab1 = "Luftfeuchtigkeit",
  ylab2 = "Feinstaub (µg/m³)",
  zlab = "Temperatur",
  legend = c("LF", "Temp", "PM10", "PM2.5"),
  colors = c("blue", "red", "darkgreen", "green")
)
}
\arguments{
\item{data}{A data frame as returned by read.madavi().}

\item{main}{Title of the plot.}

\item{start}{Start time in POSIXct format,
      or as a character string in "\%Y-\%m-\%d \%H:\%M:\%S" format. If missing all data is plotted.}

\item{end}{End time in POSIXct format,
      or as a character string in "\%Y-\%m-\%d \%H:\%M:\%S" format. If missing all data is plotted.}

\item{sub}{Subtitle}

\item{xlab}{A title for the x axis (time).}

\item{ylab1}{A title for the upper y axis (humidity).}

\item{ylab2}{A title for the lower y axis (fine dust).}

\item{zlab}{A title for the upper z axis (temperature).}

\item{legend}{Text for the plot legend describing humidity, temperature, PM10, and PM2.5.}

\item{colors}{Four color definitions used to plot humidity, temperature, PM10, and PM2.5,
      respectively.}
}
\description{
Plots objects of class airData.
}
\examples{
\dontrun{
x <- read.madavi(file.path("~","fine_dust_data"))
plot.airData(
  x,
  start="2019-12-10 09:00:00",
  end="2019-12-10 22:00:00"
)
}
}
