% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.madavi.R
\name{read.madavi}
\alias{read.madavi}
\title{Import CSV data as provided by madavi.de}
\usage{
read.madavi(path, dropEmptyCols = TRUE, na.rm = TRUE, tz = "Europe/Berlin")
}
\arguments{
\item{path}{Either a path to a single CSV file,
      or to a directory which is then being scanned for CSV files.}

\item{dropEmptyCols}{Logical, whether to drop all columns without any data.}

\item{na.rm}{Logical,
      whether to drop any row with missing data in any of the core data columns,
"PM10", "PM2_5", "humidity", or "temperature".}

\item{tz}{Time zone, used to calculate local time from the UTC raw data.}
}
\value{
An object of class \code{\link[paRticulates:airData-class]{airData}}.
}
\description{
Import CSV data as provided by madavi.de
}
\examples{
\dontrun{
# import single CSV file ~/fine_dust_data/data-esp8266-1234567-2019-12-10.csv
x <- read.madavi(file.path("~","fine_dust_data","data-esp8266-1234567-2019-12-10.csv"))

# import all CSV files from directory ~/fine_dust_data
x <- read.madavi(file.path("~","fine_dust_data"))
}
}
