% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAQI_sub.R
\name{EAQI_sub}
\alias{EAQI_sub}
\title{Calcualte particulate matter sub-index of the European Air Quality Index (EAQI)}
\usage{
EAQI_sub(x, norm = "PM10", latex_cellcolors = TRUE)
}
\arguments{
\item{x}{A numeric vector of particulate matter measurements (µg/m³).}

\item{norm}{One value of either "PM10" or "PM2.5",
      setting the PM size of the given values of \code{x}.}

\item{latex_cellcolors}{Logical,
      if true the "EAQI_color_*" columns will show LaTeX \code{\\cellcolor} code instead
of HTML colors in a \code{<span>}. See \code{Colors} section.}
}
\value{
If \code{raw=TRUE}, a numeric vector with the raw EAQI values. Otherwise a data frame with
   four columns: raw input data, EAQI level (factor), recommended color code (factor).
   The column names start with "raw_", "EAQI_level_", and "EAQI_color_",
      and end with either "PM10" or
   "PM2_5", depending on the value of \code{norm}.
}
\description{
The function returns the respective hourly EAQI grading.
}
\details{
The index has six levels, named "good" (best air quality), "fair", "moderate", "poor",
      "very poor",
and "extremely poor" (worst air quality).
Thresholds are different between PM10 and PM2.5 particles.
}
\section{Colors}{
 The data frame contains rows to support coloring tables or plots,
      e.g. if you use the data in
   an RMarkdown document. If \code{latex_cellcolors=FALSE},
      they provide a \code{<span>} with hexadecimal HTML color codes
   for each index value. Otherwise,
      the factor labels are in LaTeX format using the \code{\\cellcolor} command.
   This is useful, e.g.,
      if combined with the \code{kableExtra::kable} function. To use the code out of the box,
   define the following colors in your preamble (example shows usage in a YAML header):
   
\preformatted{
header-includes:
   - \definecolor{EAQIep}{HTML}{7D2181}
   - \definecolor{EAQIvp}{HTML}{960032}
   - \definecolor{EAQIp}{HTML}{FF5050}
   - \definecolor{EAQIm}{HTML}{F0E641}
   - \definecolor{EAQIf}{HTML}{50CCAA}
   - \definecolor{EAQIg}{HTML}{50F0E6}
}
}

\examples{
EAQI_sub(55)
}
